requirements.txt
certifi==2018.1.18
chardet==3.0.4
confuse==0.4.0
idna==2.6
oauthlib==2.0.7
pushbullet.py==0.11.0
python-dateutil==2.7.2
python-magic==0.4.15
pyyaml==3.12
requests-oauthlib==0.8.0
requests==2.18.4
six==1.11.0
twitterapi==2.5.0
urllib3==1.22
websocket-client==0.47.0
